requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
dbfread>=2.0.7
reportlab>=4.0.0
Pillow>=10.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
rarfile>=4.0